requirements.txt
numpy
moviepy
Pillow
customtkinter
packaging